+++ addons/confirma/src/classes/TestingMethod.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Reflection;
 using Confirma.Attributes;
+using Confirma.Classes.Discovery;
 using Confirma.Exceptions;
 using Confirma.Helpers;
 using Confirma.Types;
@@ -35,7 +36,7 @@ public class TestingMethod
                 {
                     Result.TestsIgnored++;
 
-                    TestLog log = new (Enums.Elogtype.Method,Name, Ignored, test.Params, attr.Reason);
+                    TestLog log = new(Enums.Elogtype.Method, Name, Ignored, test.Params, attr.Reason);
                     Result.TestLogs.Add(log);
                     continue;
                 }
@@ -45,14 +46,14 @@ public class TestingMethod
                     test.Run();
                     Result.TestsPassed++;
 
-                    TestLog log = new(Enums.Elogtype.Method ,Name, Passed, test.Params);
+                    TestLog log = new(Enums.Elogtype.Method, Name, Passed, test.Params);
                     Result.TestLogs.Add(log);
                 }
                 catch (ConfirmAssertException e)
                 {
                     Result.TestsFailed++;
 
-                    TestLog log = new(Enums.Elogtype.Method,Name, Failed ,test.Params, e.Message);
+                    TestLog log = new(Enums.Elogtype.Method, Name, Failed, test.Params, e.Message);
                     Result.TestLogs.Add(log);
 
                     if (test.Repeat?.FailFast == true)
@@ -74,7 +75,7 @@ public class TestingMethod
     private List<TestCase> DiscoverTestCases()
     {
         List<TestCase> cases = new();
-        using IEnumerator<System.Attribute> discovered = TestDiscovery
+        using IEnumerator<System.Attribute> discovered = CsTestDiscovery
             .GetTestCasesFromMethod(Method)
             .GetEnumerator();