+++ addons/confirma/src/types/TestsProps.cs
@@ -1,4 +1,5 @@
 using System;
+using Confirma.Scenes;
 
 namespace Confirma.Types;
 
@@ -7,16 +8,20 @@ public struct TestsProps
     public event Action? ExitOnFailure;
 
     public TestResult Result { get; set; } = new();
+    public ConfirmaAutoload? Autoload { get; set; }
 
     public bool RunTests { get; set; }
     public bool IsVerbose { get; set; }
     public bool IsHeadless { get; set; }
     public bool ExitOnFail { get; set; }
+    public bool DisableCsharp { get; set; }
+    public bool MonitorOrphans { get; set; }
     public bool QuitAfterTests { get; set; }
+    public bool DisableGdScript { get; set; }
+    public bool DisableParallelization { get; set; }
     public string ClassName { get; set; } = string.Empty;
     public string MethodName { get; set; } = string.Empty;
-    public bool DisableParallelization { get; set; }
-    public bool MonitorOrphans { get; set; }
+    public string GdTestPath { get; set; } = "./gdtests";
 
     public TestsProps() { }